WORKS! need to flip z, y axis around. in addition, y might be negative compared to what we see.

+++ src/Color.h
@@ -2,9 +2,9 @@
 
 struct Color
 {
-  uint16_t r;
-  uint16_t g;
-  uint16_t b;     
+  uint8_t r;
+  uint8_t g;
+  uint8_t b;     
   
   Color() :
   r(0), g(0), b(0)
